// src/generated/schemas.ts
export type Classified = {
  contact?: ClassifiedContact | undefined;
  currency: ClassifiedCurrency;
  description?: string | undefined;
  energy?: Energy | undefined;
  id: string;
  location: Location;
  media?: Media | undefined;
  meta: ClassifiedMeta;
  property: PropertyAttributes;
  propertyId: string;
  publishedAt?: (string | null) | undefined;
  publisher: ClassifiedPublisher;
  source: ClassifiedSource;
  squareUnit: ClassifiedSquareUnit;
  status: ClassifiedStatus;
  title?: string | undefined;
  transaction: Transaction;
};
export type ClassifiedContact = Partial<{
  email: string;
  name: string;
  phone: string;
}>;
export type ClassifiedCurrency = 'euro';
export type Energy = Partial<{
  dpe: EnergyDpe;
  ges: EnergyGes;
  heating: EnergyHeating;
}>;
export type EnergyDpe = Partial<{
  date: string;
  label: EnergyDpeLabel;
  value: number;
}>;
export type EnergyDpeLabel = 'a' | 'b' | 'c' | 'd' | 'e' | 'f' | 'g';
export type EnergyGes = Partial<{
  date: string;
  label: EnergyGesLabel;
  value: number;
}>;
export type EnergyGesLabel = 'a' | 'b' | 'c' | 'd' | 'e' | 'f' | 'g';
export type EnergyHeating = Partial<{
  source:
    | 'aerothermal'
    | 'biomass'
    | 'coal'
    | 'electricity'
    | 'fuel'
    | 'gas'
    | 'geothermal'
    | 'pellet'
    | 'solar'
    | 'wood';
  system: 'fireplace' | 'floor' | 'radiator' | 'reversibleAirConditioning' | 'stove';
  type: 'city' | 'individual' | 'mixed' | 'shared';
}>;
export type Location = {
  city: LocationCity;
  country: 'france';
  department: string;
  latitude?: number | undefined;
  longitude?: number | undefined;
  postcode: string;
};
export type LocationCity = {
  inseeCode: string;
  name: string;
};
export type Media = {
  images: Array<MediaImage>;
};
export type MediaImage = {
  averageHash?: string | undefined;
  differenceHash?: string | undefined;
  id: string;
  perceptualHash?: string | undefined;
  position: number;
  url: string;
};
export type ClassifiedMeta = {
  firstSeenAt: string;
  lastModifiedAt: string;
  lastSeenAt: string;
  removedAt?: (string | null) | undefined;
};
export type PropertyAttributes = {
  airConditioningExists?: boolean | undefined;
  alarmExists?: boolean | undefined;
  area?: number | undefined;
  atticExists?: boolean | undefined;
  balconyArea?: number | undefined;
  balconyCount?: number | undefined;
  balconyExists?: boolean | undefined;
  bathroomCount?: number | undefined;
  bedroomCount?: number | undefined;
  caretakerExists?: boolean | undefined;
  cellarExists?: boolean | undefined;
  constructionYear?: number | undefined;
  doorCodeExists?: boolean | undefined;
  elevatorExists?: boolean | undefined;
  fireplaceExists?: boolean | undefined;
  floor?: number | undefined;
  floorCount?: number | undefined;
  garageExists?: boolean | undefined;
  gardenExists?: boolean | undefined;
  intercomExists?: boolean | undefined;
  isCondominium?: boolean | undefined;
  isDisabledFriendly?: boolean | undefined;
  isDualAspect?: boolean | undefined;
  isNew?: boolean | undefined;
  isOccupied?: boolean | undefined;
  landArea?: number | undefined;
  livingRoomArea?: number | undefined;
  orientation?:
    | ('east' | 'north' | 'northEast' | 'northWest' | 'south' | 'southEast' | 'southWest' | 'west')
    | undefined;
  parkingExists?: boolean | undefined;
  renovationIsNeeded?: boolean | undefined;
  renovationYear?: number | undefined;
  roomCount?: number | undefined;
  showerRoomCount?: number | undefined;
  swimmingPoolExists?: boolean | undefined;
  terraceArea?: number | undefined;
  terraceCount?: number | undefined;
  terraceExists?: boolean | undefined;
  toiletCount?: number | undefined;
  toiletSeparateExists?: boolean | undefined;
  type: PropertyType;
};
export type PropertyType =
  | 'apartment'
  | 'building'
  | 'house'
  | 'land'
  | 'office'
  | 'other'
  | 'parking'
  | 'premises'
  | 'program'
  | 'shop';
export type ClassifiedPublisher = {
  email?: string | undefined;
  feesUrl?: string | undefined;
  isProfessional: boolean;
  phone?: string | undefined;
  siren?: string | undefined;
  siret?: string | undefined;
  type?: ('agency' | 'mandatary') | undefined;
};
export type ClassifiedSource = {
  domain: string;
  url: string;
};
export type ClassifiedSquareUnit = 'squareMeter';
export type ClassifiedStatus = {
  current: ClassifiedStatusValue;
};
export type ClassifiedStatusValue = 'available' | 'removed' | 'rented' | 'sold' | 'underOffer';
export type Transaction = {
  price: TransactionPrice;
  type: TransactionType;
};
export type TransactionPrice = {
  current: number;
  history: Array<TransactionPriceHistoryEntry>;
  initial: number;
  perSquareUnit?: number | undefined;
};
export type TransactionPriceHistoryEntry = {
  id: string;
  timestamp: string;
  value: number;
};
export type TransactionType = 'rent' | 'sale';
// ...
export const ClassifiedContact: z.ZodType<ClassifiedContact> = z
  .object({
    email: z.string().email(),
    name: z.string(),
    phone: z.string().regex(/^\+?[1-9]\d{1,14}$/),
  })
  .partial()
  .passthrough();
export const ClassifiedCurrency = z.literal('euro');
export const EnergyDpeLabel = z.enum(['a', 'b', 'c', 'd', 'e', 'f', 'g']);
export const EnergyDpe: z.ZodType<EnergyDpe> = z
  .object({ date: z.string(), label: EnergyDpeLabel, value: z.number().int() })
  .partial()
  .passthrough();
export const EnergyGesLabel = z.enum(['a', 'b', 'c', 'd', 'e', 'f', 'g']);
export const EnergyGes: z.ZodType<EnergyGes> = z
  .object({ date: z.string(), label: EnergyGesLabel, value: z.number().int() })
  .partial()
  .passthrough();
export const EnergyHeating: z.ZodType<EnergyHeating> = z
  .object({
    source: z.enum([
      'aerothermal',
      'biomass',
      'coal',
      'electricity',
      'fuel',
      'gas',
      'geothermal',
      'pellet',
      'solar',
      'wood',
    ]),
    system: z.enum(['fireplace', 'floor', 'radiator', 'reversibleAirConditioning', 'stove']),
    type: z.enum(['city', 'individual', 'mixed', 'shared']),
  })
  .partial()
  .passthrough();
export const Energy: z.ZodType<Energy> = z
  .object({ dpe: EnergyDpe, ges: EnergyGes, heating: EnergyHeating })
  .partial()
  .passthrough();
export const LocationCity: z.ZodType<LocationCity> = z
  .object({ inseeCode: z.string(), name: z.string() })
  .passthrough();
export const Location: z.ZodType<Location> = z
  .object({
    city: LocationCity,
    country: z.literal('france'),
    department: z.string(),
    latitude: z.number().gte(-90).lte(90).optional(),
    longitude: z.number().gte(-180).lte(180).optional(),
    postcode: z.string(),
  })
  .passthrough();
export const MediaImage: z.ZodType<MediaImage> = z
  .object({
    averageHash: z.string().optional(),
    differenceHash: z.string().optional(),
    id: z.string().uuid(),
    perceptualHash: z.string().optional(),
    position: z.number().int(),
    url: z.string().url(),
  })
  .passthrough();
export const Media: z.ZodType<Media> = z.object({ images: z.array(MediaImage) }).passthrough();
export const ClassifiedMeta: z.ZodType<ClassifiedMeta> = z
  .object({
    firstSeenAt: z.string().datetime({ offset: true }),
    lastModifiedAt: z.string().datetime({ offset: true }),
    lastSeenAt: z.string().datetime({ offset: true }),
    removedAt: z.string().datetime({ offset: true }).nullish(),
  })
  .passthrough();
export const PropertyType = z.enum([
  'apartment',
  'building',
  'house',
  'land',
  'office',
  'other',
  'parking',
  'premises',
  'program',
  'shop',
]);
export const PropertyAttributes: z.ZodType<PropertyAttributes> = z
  .object({
    airConditioningExists: z.boolean().optional(),
    alarmExists: z.boolean().optional(),
    area: z.number().gte(0).optional(),
    atticExists: z.boolean().optional(),
    balconyArea: z.number().gte(0).optional(),
    balconyCount: z.number().int().gte(0).optional(),
    balconyExists: z.boolean().optional(),
    bathroomCount: z.number().int().gte(0).optional(),
    bedroomCount: z.number().int().gte(0).optional(),
    caretakerExists: z.boolean().optional(),
    cellarExists: z.boolean().optional(),
    constructionYear: z.number().int().gte(0).optional(),
    doorCodeExists: z.boolean().optional(),
    elevatorExists: z.boolean().optional(),
    fireplaceExists: z.boolean().optional(),
    floor: z.number().int().gte(0).optional(),
    floorCount: z.number().int().gte(0).optional(),
    garageExists: z.boolean().optional(),
    gardenExists: z.boolean().optional(),
    intercomExists: z.boolean().optional(),
    isCondominium: z.boolean().optional(),
    isDisabledFriendly: z.boolean().optional(),
    isDualAspect: z.boolean().optional(),
    isNew: z.boolean().optional(),
    isOccupied: z.boolean().optional(),
    landArea: z.number().gte(0).optional(),
    livingRoomArea: z.number().gte(0).optional(),
    orientation: z
      .enum(['east', 'north', 'northEast', 'northWest', 'south', 'southEast', 'southWest', 'west'])
      .optional(),
    parkingExists: z.boolean().optional(),
    renovationIsNeeded: z.boolean().optional(),
    renovationYear: z.number().int().gte(0).optional(),
    roomCount: z.number().int().gte(0).optional(),
    showerRoomCount: z.number().int().gte(0).optional(),
    swimmingPoolExists: z.boolean().optional(),
    terraceArea: z.number().gte(0).optional(),
    terraceCount: z.number().int().gte(0).optional(),
    terraceExists: z.boolean().optional(),
    toiletCount: z.number().int().gte(0).optional(),
    toiletSeparateExists: z.boolean().optional(),
    type: PropertyType,
  })
  .passthrough();
export const ClassifiedPublisher: z.ZodType<ClassifiedPublisher> = z
  .object({
    email: z.string().email().optional(),
    feesUrl: z.string().url().optional(),
    isProfessional: z.boolean(),
    phone: z
      .string()
      .regex(/^\+?[1-9]\d{1,14}$/)
      .optional(),
    siren: z.string().optional(),
    siret: z.string().optional(),
    type: z.enum(['agency', 'mandatary']).optional(),
  })
  .passthrough();
export const ClassifiedSource: z.ZodType<ClassifiedSource> = z
  .object({ domain: z.string(), url: z.string().url() })
  .passthrough();
export const ClassifiedSquareUnit = z.literal('squareMeter');
export const ClassifiedStatusValue = z.enum([
  'available',
  'removed',
  'rented',
  'sold',
  'underOffer',
]);
export const ClassifiedStatus: z.ZodType<ClassifiedStatus> = z
  .object({ current: ClassifiedStatusValue })
  .passthrough();
export const TransactionPriceHistoryEntry: z.ZodType<TransactionPriceHistoryEntry> = z
  .object({
    id: z.string().uuid(),
    timestamp: z.string().datetime({ offset: true }),
    value: z.number().int().gte(0),
  })
  .passthrough();
export const TransactionPrice: z.ZodType<TransactionPrice> = z
  .object({
    current: z.number().int().gte(0),
    history: z.array(TransactionPriceHistoryEntry),
    initial: z.number().int().gte(0),
    perSquareUnit: z.number().int().gte(0).optional(),
  })
  .passthrough();
export const TransactionType = z.enum(['rent', 'sale']);
export const Transaction: z.ZodType<Transaction> = z
  .object({ price: TransactionPrice, type: TransactionType })
  .passthrough();
export const Classified: z.ZodType<Classified> = z
  .object({
    contact: ClassifiedContact.optional(),
    currency: ClassifiedCurrency,
    description: z.string().optional(),
    energy: Energy.optional(),
    id: z.string().uuid(),
    location: Location,
    media: Media.optional(),
    meta: ClassifiedMeta,
    property: PropertyAttributes,
    propertyId: z.string().uuid(),
    publishedAt: z.string().datetime({ offset: true }).nullish(),
    publisher: ClassifiedPublisher,
    source: ClassifiedSource,
    squareUnit: ClassifiedSquareUnit,
    status: ClassifiedStatus,
    title: z.string().optional(),
    transaction: Transaction,
  })
  .passthrough();